+++ test/function_test.py
@@ -2,6 +2,36 @@
 """
 >>> import random
 >>> from samila.functions import *
+>>> is_valid_color("blue")
+True
+>>> is_valid_color((0,0,0))
+True
+>>> is_valid_color((0.1,0.1,0,1))
+True
+>>> is_valid_color([1,1,1,1])
+True
+>>> is_valid_color([1,1,1,1,1,1])
+False
+>>> is_valid_color("nothing")
+False
+>>> is_valid_color("#FFFAAF")
+True
+>>> color_complement("#FFFFFF")
+'#000000'
+>>> color_complement("#FFAFBF")
+'#005040'
+>>> color_complement("#000000")
+'#FFFFFF'
+>>> select_color("blue")
+'blue'
+>>> select_color("#FFFFFA")
+'#FFFFFA'
+>>> select_color((0.1,0.1,0.1))
+(0.1, 0.1, 0.1)
+>>> select_color(2)
+>>> select_color(None)
+>>> select_color("complement")
+'COMPLEMENT'
 >>> s = list(float_range(1,1.5,0.1))
 >>> s
 [1.0, 1.1, 1.2000000000000002, 1.3000000000000003, 1.4000000000000004]
@@ -13,16 +43,16 @@ False
 False
 >>> is_same_data(s,[])
 False
->>> filter_color("yellow")
-'yellow'
->>> filter_color((0.2,0.3,0.4))
-(0.2, 0.3, 0.4)
->>> filter_color("#FFFFFF")
-'#FFFFFF'
+>>> filter_color("yellow", "blue")
+('yellow', 'blue')
+>>> filter_color((0.2,0.3,0.4), (0.2,0.3,0.4,1))
+((0.2, 0.3, 0.4), (0.2, 0.3, 0.4, 1))
+>>> filter_color("#FFFFFF", "#ffffe1")
+('#FFFFFF', '#ffffe1')
 >>> random.seed(2)
->>> color1 = filter_color("random")
+>>> color1, bgcolor1 = filter_color("random", "random")
 >>> random.seed(3)
->>> color2 = filter_color("RANDOM")
+>>> color2, bgcolor2 = filter_color("RANDOM", "RANDOM")
 >>> color1 == color2
 False
 >>> random.seed(2)
@@ -35,8 +65,10 @@ False
 7
 >>> len(color2)
 7
->>> filter_color(2)
->>> filter_color(4)
+>>> filter_color(2,2)
+(None, None)
+>>> filter_color(4,3)
+(None, None)
 >>> filter_size(2)
 >>> filter_size((2, 'test'))
 >>> filter_size((2, 3.5))